New Wave Prefab, altered Enemy script, new PathDriver script, updated Water & Wave scripts, new tags, and minor changes
- New Wave Prefab
Deleted the old mesh based wave. Created a new one with the empty box zone. This moves along a spline to update specific vertices on the water mesh.
- Altered Enemy script.
Took out the movement and path-based code, creating a new parent class that Enemy derives from.
- New PathDriver script.
This took all the movement and path-based code from the Enemy script.
- Updated Water script.
Removed unnecessary code. Made some variables private. No longer adds component, instead marking it as a required component upon adding it to a GameObject.  Cleared the mesh variable during a function.
- Updated Wave script.
Added compatibility with Unity Splines. Removed unnecessary variables, added a couple of new ones, and changed some to private. Since this is now a prefab, it needed to retrieve some references during runtime. New variables to adjust the speed of the rise and decrease of the water vertices. New for loop to lower water vertices if above the default height. Clearing the localVertexIndicies in FindLocalVertices().
- New tags to define spline game objects for finding during runtime.
- Minor changes include new material and removal of WaterGrid test script.