Instances update
Scripts included in this are the ones exclusively affected (Other scripts not included in this push had additional changes).

This push changes from the old system of assigning a script to a variable for it's functions to be called. Instead now using instancing/singletons(?) to call without extra variables n stuff.